Explicitly specify a temporary directory
It is common in restricted enterprise setups that the /tmp directory will
be a tmpfs mountpoint mounted with noexec.

In those situations, sbt will fail because the JVM will try to mmap files
from its temporary location with executable permissions.

Since remounting /tmp may break security policies, we are better off leaving
that untouched, and specifying a tmpdir ourselves.