Expand ttlTolerance to apply to missing rows
It is entirely possible that a row in the target table could not exist
due to its TTL expiring within the specified tolerance.
To handle this, the ttlTolerance is checked in the case where a target
rows is missing, and no error is indicated if the source row's TTL is
within the tolerance.