Remove unused configuration properties and fix documentation
The properties `scanSegments` and `maxMapTasks` of the `target` table are never used by the DynamoDB connector.

I also fixed the documentation, which was documenting the `readThroughput` of the `target` table instead of the `writeThroughput`.

Finally, I fixed the default throughput calculator to use RCU and WCU as units instead of bytes per second, because the DynamoDB connector expects RCU and WCU.